robot control/DOF/diffused_fields/requirements.txt
# Core dependencies for diffused_fields package
numpy>=1.21.0
scipy>=1.7.0
PyYAML>=6.0

# Point cloud processing
open3d>=0.13.0
potpourri3d
robust_laplacian
pcdiff

# Visualization
polyscope
matplotlib>=3.5.0